Split UI app state into feature modules
