+++ apps/ui/src/app/demo_mode.ts
@@ -1,6 +1,7 @@
 import { defaultLiveAnalysisConfig } from "../constants";
 import { EXPECTED_LIVE_PAYLOAD_SCHEMA_VERSION } from "../transport/live_models";
-import { composeVehicleSettings, type AppState } from "./ui_app_state";
+import { composeVehicleSettings } from "./settings_state";
+import type { AppState } from "./ui_app_state";
 import { batch } from "./ui_signals";
 
 type DemoDeps = {
