+++ apps/ui/src/app/features/realtime_feature.ts
@@ -1,12 +1,10 @@
 import type { QueryClient } from "@tanstack/query-core";
 
 import type { FeatureFormatting, FeatureServices } from "../feature_deps_base";
-import type {
-  RealtimeState,
-  SettingsState,
-  ShellState,
-  SpectrumState,
-} from "../ui_app_state";
+import type { RealtimeState } from "../realtime_state";
+import type { SettingsState } from "../settings_state";
+import type { ShellState } from "../shell_state";
+import type { SpectrumState } from "../spectrum_state";
 import type { RealtimeLoggingPanelBridge } from "../views/realtime_logging_panel";
 import type { RealtimeLiveOverviewBridge } from "../views/realtime_live_overview";
 import type { SensorsPanelView } from "../views/sensors_panel";
