+++ apps/ui/src/app/features/realtime_feature_view_state.ts
@@ -4,12 +4,10 @@ import type {
   RealtimeFeatureWorkflowLoggingError,
   RealtimeFeatureWorkflowSignals,
 } from "./realtime_feature_workflow";
-import type {
-  RealtimeState,
-  SettingsState,
-  ShellState,
-  SpectrumState,
-} from "../ui_app_state";
+import type { RealtimeState } from "../realtime_state";
+import type { SettingsState } from "../settings_state";
+import type { ShellState } from "../shell_state";
+import type { SpectrumState } from "../spectrum_state";
 import {
   bindReplaceableTimerEffect,
   createReplaceableInterval,
