+++ apps/ui/src/app/signal_state.ts
@@ -1,0 +1,5 @@
+import type { Signal } from "./ui_signals";
+
+export type SignalState<T extends object> = {
+  [K in keyof T]: Signal<T[K]>;
+};
