+++ apps/ui/tests/typecheck-baseline.txt
@@ -64,16 +64,16 @@ tests/spectrum_css_vars.spec.ts(90,7): error TS2349: This expression is not call
 tests/spectrum_test_support.ts(160,5): error TS2352: Conversion of type '{ getPropertyValue: () => string; }' to type 'CSSStyleDeclaration' may be a mistake because neither type sufficiently overlaps with the other. If this was intentional, convert the expression to 'unknown' first.
 tests/spectrum_test_support.ts(59,14): error TS2339: Property 'insertBefore' does not exist on type '{}'.
 tests/spectrum_test_support.ts(63,19): error TS2339: Property 'insertBefore' does not exist on type '{}'.
-tests/ui_live_payload_application.spec.ts(101,19): error TS2741: Property 'frame_samples' is missing in type '{ id: string; name: string; connected: boolean; mac_address: string; location_code: string; last_seen_age_ms: number; dropped_frames: number; frames_total: number; sample_rate_hz: number; firmware_version: string; }' but required in type 'AdaptedClient'.
-tests/ui_live_payload_application.spec.ts(148,19): error TS2741: Property 'frame_samples' is missing in type '{ id: string; name: string; connected: boolean; mac_address: string; location_code: string; last_seen_age_ms: number; dropped_frames: number; frames_total: number; sample_rate_hz: number; firmware_version: string; }' but required in type 'AdaptedClient'.
-tests/ui_live_payload_application.spec.ts(220,19): error TS2741: Property 'frame_samples' is missing in type '{ id: string; name: string; connected: boolean; mac_address: string; location_code: string; last_seen_age_ms: number; dropped_frames: number; frames_total: number; sample_rate_hz: number; firmware_version: string; }' but required in type 'AdaptedClient'.
-tests/ui_live_payload_application.spec.ts(43,19): error TS2741: Property 'frame_samples' is missing in type '{ id: string; name: string; connected: boolean; mac_address: string; location_code: string; last_seen_age_ms: number; dropped_frames: number; frames_total: number; sample_rate_hz: number; firmware_version: string; }' but required in type 'AdaptedClient'.
+tests/ui_live_payload_application.spec.ts(146,19): error TS2741: Property 'frame_samples' is missing in type '{ id: string; name: string; connected: boolean; mac_address: string; location_code: string; last_seen_age_ms: number; dropped_frames: number; frames_total: number; sample_rate_hz: number; firmware_version: string; }' but required in type 'AdaptedClient'.
+tests/ui_live_payload_application.spec.ts(218,19): error TS2741: Property 'frame_samples' is missing in type '{ id: string; name: string; connected: boolean; mac_address: string; location_code: string; last_seen_age_ms: number; dropped_frames: number; frames_total: number; sample_rate_hz: number; firmware_version: string; }' but required in type 'AdaptedClient'.
+tests/ui_live_payload_application.spec.ts(41,19): error TS2741: Property 'frame_samples' is missing in type '{ id: string; name: string; connected: boolean; mac_address: string; location_code: string; last_seen_age_ms: number; dropped_frames: number; frames_total: number; sample_rate_hz: number; firmware_version: string; }' but required in type 'AdaptedClient'.
+tests/ui_live_payload_application.spec.ts(99,19): error TS2741: Property 'frame_samples' is missing in type '{ id: string; name: string; connected: boolean; mac_address: string; location_code: string; last_seen_age_ms: number; dropped_frames: number; frames_total: number; sample_rate_hz: number; firmware_version: string; }' but required in type 'AdaptedClient'.
 tests/ui_shell_runtime_modules.spec.ts(122,5): error TS2349: This expression is not callable.
 tests/ui_shell_runtime_modules.spec.ts(32,68): error TS2352: Conversion of type '{ clearTimeout: () => undefined; setTimeout: Window["setTimeout"]; }' to type 'Window & typeof globalThis' may be a mistake because neither type sufficiently overlaps with the other. If this was intentional, convert the expression to 'unknown' first.
 tests/ui_shell_runtime_modules.spec.ts(397,5): error TS2349: This expression is not callable.
 tests/ui_shell_runtime_modules.spec.ts(451,5): error TS2349: This expression is not callable.
-tests/ui_spectrum_controller.spec.ts(350,47): error TS2749: 'UiSpectrumController' refers to a value, but is being used as a type here. Did you mean 'typeof UiSpectrumController'?
-tests/ui_spectrum_controller.spec.ts(353,51): error TS7006: Parameter 'prepared' implicitly has an 'any' type.
+tests/ui_spectrum_controller.spec.ts(348,47): error TS2749: 'UiSpectrumController' refers to a value, but is being used as a type here. Did you mean 'typeof UiSpectrumController'?
+tests/ui_spectrum_controller.spec.ts(351,51): error TS7006: Parameter 'prepared' implicitly has an 'any' type.
 tests/ui_startup_coordinator.spec.ts(82,3): error TS2322: Type 'Location' is not assignable to type 'string & Location'.
 tests/ui_startup_coordinator.spec.ts(89,14): error TS2790: The operand of a 'delete' operator must be optional.
 tests/ui_startup_coordinator.spec.ts(92,5): error TS2322: Type 'Location' is not assignable to type 'string & Location'.
