+++ docs/ai/repo-map.md
@@ -48,6 +48,7 @@ Paths below are repo-relative unless a line explicitly calls out a Python import
 - `apps/server/vibesensor/shared/`: cross-cutting ports, model/payload types, JSON helpers, boundary codecs, split constant modules under `apps/server/vibesensor/shared/constants/`, and small package-level helpers such as `apps/server/vibesensor/shared/_data_files.py`, `apps/server/vibesensor/shared/sensor_units.py`, and `apps/server/vibesensor/shared/run_context_warning.py`. Key stable owners include `apps/server/vibesensor/shared/types/persisted_analysis.py`, `apps/server/vibesensor/shared/types/analysis_views.py`, `apps/server/vibesensor/shared/types/history_analysis_contracts.py`, `apps/server/vibesensor/shared/types/run_schema.py`, `apps/server/vibesensor/shared/types/car_config.py`, `apps/server/vibesensor/shared/types/speed_source_config.py`, and the codec/projection modules under `apps/server/vibesensor/shared/boundaries/` such as `apps/server/vibesensor/shared/boundaries/settings/snapshot.py`.
 - `apps/server/vibesensor/domain/`: domain model package for classification, ranking, lifecycle, and query logic. See `docs/domain-model.md` for the domain object graph.
 - `apps/ui/src/app/runtime/`: UI composition root and runtime controllers.
+- `apps/ui/src/app/shell_state.ts`, `apps/ui/src/app/transport_state.ts`, `apps/ui/src/app/realtime_state.ts`, `apps/ui/src/app/history_state.ts`, `apps/ui/src/app/settings_state.ts`, and `apps/ui/src/app/spectrum_state.ts`: focused owners for top-level AppState slice types, defaults, factories, and pure update helpers; `apps/ui/src/app/ui_app_state.ts` stays the thin composition/compatibility surface.
 - `apps/ui/src/app/features/`: UI feature workflows for settings, realtime, history, updates, cars, and ESP flash, plus the shared polling controller used by long-running status features.
 - `apps/ui/src/app/views/`: DOM renderers and event-target decoders, including the car wizard view helpers.
 - `apps/ui/src/config.ts`: centralized UI tuning constants (polling intervals, spectrum bounds, history heatmap positions).
